app/requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0
psycopg2-binary==2.9.7
mlflow==2.5.0
joblib==1.3.1